Move each filter to its own file
